requirements.txt
Pillow
EasyDict
opencv-python
shapely
Cython
scipy
pandas
pyyaml
json_tricks
scikit-image
yacs
imageio-ffmpeg
h5py
tensorboard
future
# Cluster needs CUDA 9.2 version so we install separately as:
# pip3 install torch==1.3.0+cu92 torchvision==0.4.1+cu92 -f https://download.pytorch.org/whl/torch_stable.html
torch
torchvision
